chore: 컨벤션 적용
