+++ app/src/main/java/com/techcourse/RegisterController.java
@@ -1,5 +1,7 @@
-package com.techcourse.controller;
+package com.techcourse;
 
+import com.techcourse.domain.User;
+import com.techcourse.repository.InMemoryUserRepository;
 import context.org.springframework.stereotype.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -9,10 +11,17 @@ import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.view.JspView;
 
 @Controller
-public class RegisterViewController{
+public class RegisterController {
 
-    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
     public ModelAndView execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        return new ModelAndView(new JspView("/register.jsp"));
+        final var user = new User(
+                2,
+                req.getParameter("account"),
+                req.getParameter("password"),
+                req.getParameter("email"));
+        InMemoryUserRepository.save(user);
+
+        return new ModelAndView(new JspView("redirect:/index.jsp"));
     }
 }
