+++ app/src/main/java/com/techcourse/UserController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
